+++ tests/testthat/test-format.R
@@ -1,0 +1,20 @@
+context("format")
+
+test_that("format", {
+  expect_error(tmr_format(1), "^`x` must be class hms[.]$")
+  expect_identical(tmr_format(tmr_timer(61.36)), "00:01:01.360")
+  expect_identical(tmr_format(tmr_timer(61.36), digits = 1), "00:01:01.4")
+  expect_identical(tmr_format(tmr_timer(61.36), digits = 0), "00:01:01")
+  expect_identical(tmr_format(tmr_timer(59.6), digits = 0), "00:01:00")
+  expect_identical(tmr_format(tmr_timer(59.6), digits = 1), "00:00:59.6")
+  expect_identical(tmr_format(tmr_timer(59.6), digits = 3), "00:00:59.600")
+  expect_identical(tmr_format(tmr_timer(59.6), digits = 6), "00:00:59.600000")
+  expect_identical(tmr_format(tmr_timer(59.60001), digits = 6), "00:00:59.600010")
+  expect_identical(tmr_format(tmr_timer(59.60001), digits = -3), "00:00:00")
+  expect_identical(tmr_format(tmr_timer(59.60001), digits = 0), "00:01:00")
+  expect_identical(tmr_format(tmr_timer(58.40001), digits = 0), "00:00:58")
+  expect_identical(tmr_format(tmr_timer(123.40001), digits = 0), "00:02:03")
+  expect_identical(tmr_format(tmr_timer(123.40001), digits = -1), "00:02:00")
+ # expect_identical(tmr_format(tmr_timer(123.40001), digits = -2), "00:02:00")
+  }
+)